fixed some build stuff for ase

+++ scripts/build_opae.sh
@@ -39,5 +39,10 @@ make install
 
 if [ "$OPENCL_ASE_SIM" == "1" ]; then
 	#copy the ASE lib over the regular lib so that we don't have to relink our program
-	cp $OPAE_LOCAL_INST_PATH/lib/libopae-c-ase.so $OPAE_LOCAL_INST_PATH/lib/libopae-c.so
+	if [ -f $OPAE_LOCAL_INST_PATH/lib/libopae-c-ase.so ]; then
+		cp $OPAE_LOCAL_INST_PATH/lib/libopae-c-ase.so $OPAE_LOCAL_INST_PATH/lib/libopae-c.so
+	fi
+	if [ -f $OPAE_LOCAL_INST_PATH/lib64/libopae-c-ase.so ]; then
+		cp $OPAE_LOCAL_INST_PATH/lib64/libopae-c-ase.so $OPAE_LOCAL_INST_PATH/lib64/libopae-c.so
+	fi
 fi
